refactor(rib): split update-group membership lifecycle
